rename media distribution graph
