+++ src/app/ranking/pages/RankingProfile.tsx
@@ -7,7 +7,7 @@ import { ContestLog, Ranking } from '../interfaces'
 import RankingApi from '../api'
 import ContestApi from '../../contest/api'
 import ContestLogsByDayGraph from '../components/graphs/ReadingActivityGraph'
-import ContestLogsByMediumGraph from '../components/graphs/ContestLogsByMediumGraph'
+import ContestLogsByMediumGraph from '../components/graphs/MediaDistributionGraph'
 import ContestLogsOverview from '../components/ContestLogsOverview'
 import { rankingsToRegistrationOverview } from '../transform/graph'
 import { Contest } from '../../contest/interfaces'
